backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
channels==4.2.2
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.2
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
parts==2.1.0
products==2.0.1
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
routing==0.2.0
signals==0.0.2
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
